patch ParseToLingo
quick patch to "fix" ParseToLingo as it creates a language for each language it parses but Lingo starts with one already so there is an unused language.

Note that this parser uses the first languages so it can't add languages to an existing set of languages; it's presumed to be an empty Lingo object.